增加mini周组件
