+++ demo/events.js
@@ -3,59 +3,17 @@ const end = current + 8 * 60 * 60 * 1000;
 
 export default [
   {
-    start: '2018-12-23 14:00', // 事件开始时间
-    end: '2018-12-26 16:00', // 事件结束时间
+    start: '2018-12-20 14:00', // 事件开始时间
+    end: '2018-12-20 16:00', // 事件结束时间
+    important: true,
     // 事件的渲染函数
     render: () => <div>12-23 14:00 ~  12-26 16:00</div>,
   },
   {
-    start: '2018-12-19 14:00', // 事件开始时间
-    end: '2018-12-19 16:00', // 事件结束时间
+    start: '2018-12-20 14:00', // 事件开始时间
+    end: '2018-12-20 16:00', // 事件结束时间
     // 事件的渲染函数
-    render: () => <div>12-19 14:00 ~  12-19 16:00</div>,
+    render: () => <div>12-20 14:00 ~  12-20 16:00</div>,
   },
-  {
-    start: '2018-12-19 14:00', // 事件开始时间
-    end: '2018-12-19 16:00', // 事件结束时间
-    // 事件的渲染函数
-    render: () => <div>12-19 14:00 ~  12-19 16:00</div>,
-  },
-  {
-    start: '2018-12-19 14:00', // 事件开始时间
-    end: '2018-12-27 16:00', // 事件结束时间
-    // 事件的渲染函数
-    render: () => <div>12-19 14:00 ~  12-27 16:00</div>,
-  },
-  {
-    start: '2018-12-17 14:00', // 事件开始时间
-    end: '2018-12-27 16:00', // 事件结束时间
-    // 事件的渲染函数
-    render: () => <div>12-17 14:00 ~  12-27 16:00</div>,
-  },
-  {
-    start: '2018-12-18 14:00', // 事件开始时间
-    end: '2018-12-27 16:00', // 事件结束时间
-    // 事件的渲染函数
-    render: () => <div>12-18 14:00 ~  12-27 16:00</div>,
-  },
-  // {
-  //   start: '2018-12-23 14:00', // 事件开始时间
-  //   end: '2018-12-27 16:00', // 事件结束时间
-  //   // 事件的渲染函数
-  //   render: () => <div>12-23 14:00 ~  12-23 16:00</div>,
-  // },
-  // {
-  //   start: '2018-12-23 14:00', // 事件开始时间
-  //   end: '2018-12-29 16:00', // 事件结束时间
-  //   // 事件的渲染函数
-  //   render: () => <div>12-23 14:00 ~  12-23 16:00</div>,
-  // },
-
-  // {
-  //   start: '2018-12-18 09:00', // 事件开始时间
-  //   end: '2018-12-18 14:00', // 事件结束时间
-  //   // 事件的渲染函数
-  //   render: () => <div>12-18 09:00 ~  12-18 14:00</div>,
-  // },
 
 ];
