+++ src/date/DateTable.jsx
@@ -1,19 +1,35 @@
 // customized rc-calendar https://github.com/react-component/calendar/blob/master/
 
 import React from 'react';
+import classnames from 'classnames';
 import DateTHead from './DateTHead';
 import DateTBody from './DateTBody';
 
 export default class DateTable extends React.Component {
+  constructor(props) {
+    super(props);
+    this.state = {
+      tableHeight: 0,
+    };
+  }
+
+  componentDidMount() {
+    this.setState({
+      tableHeight: this.fullTable.offsetHeight,
+    });
+  }
+
   renderEvents() {
     const {
-      scheduleRender, startHour, step, endHour, value, type, width,
+      scheduleRender, startHour, step, endHour, value, type,
     } = this.props;
+    const { tableHeight } = this.state;
+
     const renderOpts = {
-      startHour, step, endHour, type, current: value, width, ...this.props,
+      startHour, step, endHour, type, current: value, ...this.props,
     };
     if (scheduleRender) {
-      const content = scheduleRender(renderOpts);
+      const content = scheduleRender(renderOpts, tableHeight);
       return <div className="events-month-wrapper">{content}</div>;
     }
     return '';
@@ -21,9 +37,16 @@ export default class DateTable extends React.Component {
 
   render() {
     const { prefixCls } = this.props;
+    const fullTableHeight = this.fullTable ? this.fullTable.offsetHeight : 0;
+    const cellHeight = 0.8 * (fullTableHeight - 32) / 6;
+    const tableCls = classnames({
+      [`${prefixCls}-table`]: true,
+      'super-mini': cellHeight <= 42 && cellHeight > 0,
+
+    });
     return (
-      <div className={`${prefixCls}-wrapper`}>
-        <table className={`${prefixCls}-table`} cellSpacing="0" role="grid">
+      <div className={tableCls} key={`${prefixCls}-date`}>
+        <table className={`${prefixCls}-table`} cellSpacing="0" role="grid" ref={(c) => { this.fullTable = c; }}>
           <DateTHead {...this.props} />
           <DateTBody {...this.props} />
         </table>
