+++ src/miniWeek/index.js
@@ -1,0 +1,8 @@
+/**
+ * 获取当前面板视图中哪些日期有日程，点击对应的日程进行渲染
+ * @param {object} props 用户传的配置，决定event如何渲染等 
+ * @param {array} events 
+ */
+function handleEvents(props, events) {
+
+}
