+++ src/style/common/FullCalendar.less
@@ -73,6 +73,7 @@
       bottom: 0;
       margin-top: 4px;
       text-align: center;
+      cursor: pointer;
       color: rgba(31, 56, 88, 0.3)
     }
     .more-icon {
@@ -83,6 +84,19 @@
       border-top: 4px solid transparent;
       border-left: 4px solid rgba(31, 56, 88, 0.3);
       border-bottom: 4px solid transparent;
+    }
+  }
+  .super-mini {
+    .more-event {
+      top: 3px;
+    }
+    .hot-icon {
+      position: absolute;
+      left: 50%;
+      width: 5px;
+      height: 5px;
+      border-radius: 5px;
+      background: @brand-primary;
     }
   }
   &-calendar-body {
