+++ src/style/common/index.less
@@ -9,6 +9,7 @@
   display: none;
 }
 
+@import "MiniWeek.less";
 @import "Calendar.less";
 @import "Picker.less";
 @import "FullCalendar.less";
